requirements.txt
flask
flask-cors
google-generativeai
numpy
python-dotenv